+++ .ci/script.sh
@@ -28,7 +28,9 @@ fstar.exe --version
 echo | $(which d8)
 
 make -C $FSTAR_HOME/ulib -j 4
-make -j 4 && make -C test all wasm external -j 4
+make -j 4 && \
+  make -j 4 -C kremlib && \
+  make -C test all wasm external -j 4
 
 # Tutorial refresh, only on master and not for pull requests
 if [[ $TRAVIS_BRANCH == "master" && $DZOMO_TOKEN != "" ]]; then