+++ kremlib/c/fstar_char.c
@@ -1,0 +1,5 @@
+#include "FStar_Char.h"
+
+FStar_Char_char FStar_Char_char_of_u32(uint32_t x) {
+  return x;
+}