+++ kremlib/c/fstar_uint32.c
@@ -1,0 +1,30 @@
+#include "FStar_UInt32.h"
+
+uint32_t FStar_UInt32_eq_mask(uint32_t x, uint32_t y) {
+  x = ~(x ^ y);
+  x &= x << 16;
+  x &= x << 8;
+  x &= x << 4;
+  x &= x << 2;
+  x &= x << 1;
+  return ((int32_t)x) >> 31;
+}
+
+uint32_t FStar_UInt32_gte_mask(uint32_t x, uint32_t y) {
+  return ~((uint32_t)(((int64_t)x - y) >> 63));
+}
+
+Prims_string FStar_UInt32_to_string(uint32_t i) {
+  char *buf = KRML_HOST_MALLOC(24);
+  snprintf(buf, 24, "%"PRIu32, i);
+  return buf;
+}
+
+uint32_t FStar_UInt32_uint_to_t(krml_checked_int_t x) {
+  /* TODO bounds check */
+  return x;
+}
+
+krml_checked_int_t FStar_UInt32_v(uint32_t x) {
+  RETURN_OR(x);
+}