+++ src/Helpers.ml
@@ -168,6 +168,16 @@ let fold_arrow ts t_ret =
 
 let is_array = function TArray _ -> true | _ -> false
 
+let is_null = function
+  | { node = EApp (
+      { node = EQualified (
+          ([ "LowStar"; "Buffer" ] | [ "C"; "Nullity" ]),
+          "null"); _ }, _); _ }
+  ->
+      true
+  | _ ->
+      false
+
 let is_uu name = KString.starts_with name "uu____"
 
 (* If [e2] is assigned into an expression of type [t], we can sometimes
@@ -285,16 +295,34 @@ let rec is_int_constant e =
 
 (* This is a conservative approximation. See C11 6.6. *)
 let rec is_initializer_constant e =
+  let is_address = function
+    | TArrow _ | TBuf _ | TArray _
+    (* See comment in test/TopLevelArray.fst *)
+    (*| TQualified (["C";"String"], "t") *)->
+        true
+    | _ ->
+        KPrint.bprintf "%a is not an initializer constant\n" pexpr e;
+        false
+  in
   is_int_constant e ||
   match e with
   | { node = EAddrOf { node = EQualified _; _ }; _ } ->
       true
-  | { node = EQualified _; typ = TArrow _ } ->
+  | { node = EQualified _; typ = t } ->
+      is_address t
+  | { node = EEnum _; _ } ->
       true
+  | { node = EString _; _ } ->
+      true
+  | { node = EFlat es; _ } ->
+      List.for_all (fun (_, e) -> is_initializer_constant e) es
   | { node = EBufCreateL (_, es); _ } ->
       List.for_all is_initializer_constant es
   | _ ->
-      false
+      let r = is_null e in
+      if not r then
+        KPrint.bprintf "%a is not an initializer constant\n" pexpr e;
+      r
 
 let assert_tlid t =
   (* We only have nominal typing for variants. *)
@@ -306,6 +334,11 @@ let assert_tbuf t =
 let assert_elid t =
   (* We only have nominal typing for variants. *)
   match t with EQualified lid -> lid | _ -> assert false
+
+let assert_tbuf_or_tarray t =
+  match t with
+  | TBuf t | TArray (t, _) -> t
+  | _ -> Warnings.fatal_error "%a is neither a tbuf or tarray\n" ptyp t
 
 
 (* Somewhat more advanced helpers *********************************************)
@@ -365,3 +398,70 @@ let rec nest_in_return_pos i typ f e =
 let nest_in_return_pos = nest_in_return_pos 0
 
 let push_ignore = nest_in_return_pos TUnit (fun _ e -> with_type TUnit (EIgnore (strip_cast e)))
+
+(* Big AST nodes *************************************************************)
+
+let mk_bufblit src_buf src_ofs dst_buf dst_ofs len =
+  (* This function is now used for copy-assignments in C and Wasm. There are
+   * some possibilities for optimization, e.g. using memset when the initial
+   * value is a byte or any 0 (in C). *)
+  let t = assert_tbuf_or_tarray src_buf.typ in
+  match len.node with
+  | EConstant (_, "1") ->
+      EBufWrite (dst_buf, dst_ofs, with_type t (EBufRead (src_buf, src_ofs)))
+  | _ ->
+      let b_src, body_src, ref_src =
+        mk_named_binding "src" (TBuf t) (EBufSub (src_buf, src_ofs))
+      in
+      let b_dst, body_dst, ref_dst =
+        mk_named_binding "dst" (TBuf t) (EBufSub (dst_buf, dst_ofs))
+      in
+      let b_len, body_len, ref_len =
+        mk_named_binding "len" uint32 len.node
+      in
+      let b_len = mark_mut b_len in
+      ELet (b_src, body_src, close_binder b_src (with_unit (
+      ELet (b_dst, body_dst, close_binder b_dst (with_unit (
+      ELet (b_len, body_len, close_binder b_len (with_unit (
+        EWhile (
+          mk_gt_zero ref_len, with_unit (
+          ESequence [ with_unit (
+            EBufWrite (
+              ref_dst,
+              mk_minus_one ref_len,
+              with_type t (EBufRead (ref_src, mk_minus_one ref_len)))); with_unit (
+            EAssign (ref_len, mk_minus_one ref_len))])))))))))))
+
+(* e1 := e2 *)
+let mk_copy_assignment (t, size) e1 e2 =
+  let assert_ro n e =
+    if not (is_readonly_c_expression e) then
+      Warnings.fatal_error "copy-assign, %s is not a readonly expression: %a" n pexpr e
+  in
+  let e1 = with_type (TBuf t) e1 in
+  begin match e2.node with
+  | EBufCreate (_, init, len) ->
+      if init.node = EAny then
+        ESequence []
+      else if snd size = "1" then
+        (* A copy-assignment with size 1 can become a single assignment. *)
+        EBufWrite (e1, zerou32, init)
+      else begin
+        assert_ro "e1" e1;
+        let b_init = fresh_binder "init" init.typ in
+        ELet (b_init, init,
+          with_unit (EFor (fresh_binder ~mut:true "i" uint32,
+            zerou32,
+            mk_lt32 (DeBruijn.lift 2 len),
+            mk_incr32,
+            let i = with_type uint32 (EBound 0) in
+            let init = with_type init.typ (EBound 1) in
+            with_unit (EBufWrite (DeBruijn.lift 2 e1, i, init)))))
+      end
+  | EBufCreateL (_, inits) ->
+      assert_ro "e1" e1;
+      ESequence (List.mapi (fun i init -> with_unit (EBufWrite (e1, mk_uint32 i, init))) inits)
+  | _ ->
+      let l = with_type uint32 (EConstant size) in
+      mk_bufblit e2 zerou32 e1 zerou32 l
+  end