+++ src/Kremlin.ml
@@ -131,13 +131,14 @@ The compiler switches turn on the following options.
   [-cc msvc] adds [%s]
   [-cc compcert] adds [%s]
 
-The [-fc89] option triggers [-fnouint128], [-fnoanonymous-unions],
-[-fnocompound-literals] and [-fc89-scope]. It also changes the invocations above
-to use [-std=c89].
+The [-fc89] option triggers [-fnoanonymous-unions], [-fnocompound-literals] and
+[-fc89-scope]. It also changes the invocations above to use [-std=c89]. Note
+that if you're using the uint128 type, you will have to manually compile this
+code with -DKRML_VERIFIED_UINT128.
 
 To debug Wasm codegen, it might be useful to trigger the same compilation path
 as Wasm, but emit C code instead. This can be achieved with [-wasm -d
-force-c,c-calls,wasm-calls -drop C,TestLib -add-include '"hack.h"' -fnouint128]
+force-c,c-calls,wasm-calls -drop C,TestLib -add-include '"hack.h"']
 where [hack.h] contains [#define WasmSupport_check_buffer_size(X)].
 
 Supported options:|}
@@ -204,7 +205,11 @@ Supported options:|}
     "-drop", Arg.String (fun s ->
       used_drop := true;
       List.iter (prepend Options.drop) (Parsers.drop s)),
-      "  do not extract Code for this module (see above)";
+      "  do not extract code for this module (see above)";
+    "-library", Arg.String (fun s ->
+      List.iter (prepend Options.library) (Parsers.drop s)),
+      "  this is a model and all functions should be made abstract";
+    "-extract-uint128", Arg.Set Options.extract_uint128, ""; (* no doc, intentional *)
     "-header", Arg.String (fun f ->
       let c = Utils.file_get_contents f in
       Options.header := fun _ _ -> c
@@ -227,8 +232,6 @@ Supported options:|}
       structures by value and use pointers instead";
     "-fnoanonymous-unions", Arg.Clear Options.anonymous_unions, "  disable C11 \
       anonymous unions";
-    "-fnouint128", Arg.Clear Options.uint128, "  don't assume a built-in type \
-      __uint128";
     "-fnocompound-literals", Arg.Unit (fun () -> Options.compound_literals := `Never),
       "  don't generate C11 compound literals";
     "-ftail-calls", Arg.Set Options.tail_calls, "  statically compile tail-calls \
@@ -237,10 +240,12 @@ Supported options:|}
       loops smaller than N";
     "-fparentheses", Arg.Set Options.parentheses, "  add unnecessary parentheses \
       to silence GCC and Clang's -Wparentheses";
+    "-fno-shadow", Arg.Set Options.no_shadow, "  add unnecessary renamings to \
+      defeat GCC and Clang's -Wshadow, as well as the various MSVC warnings";
     "-fcurly-braces", Arg.Set Options.curly_braces, "  always add curly braces \
       around blocks";
-    "-fshort-enums", Arg.Set Options.short_enums, "  use C macros and uint8_t \
-      for enums instead of C11 enum";
+    "-fnoshort-enums", Arg.Clear Options.short_enums, "  use C11 enums instead \
+      of C macros and uint8_t for enums";
     "-fc89-scope", Arg.Set Options.c89_scope, "  use C89 scoping rules";
     "-fc89", Arg.Set arg_c89, "  generate C89-compatible code (meta-option, see above)";
     "", Arg.Unit (fun _ -> ()), " ";
@@ -297,7 +302,6 @@ Supported options:|}
   (* Meta-options that enable other options. Do this now because it influences
    * the default options for each compiler. *)
   if !Options.wasm then begin
-    Options.uint128 := false;
     Options.anonymous_unions := false;
     (* This forces the evaluation of compound literals to happen first, before
      * they are assigned. This is the C semantics, and the C compiler will do it
@@ -306,17 +310,11 @@ Supported options:|}
     Options.struct_passing := false
   end;
 
-  if not !Options.wasm || Options.debug "force-c" then
-    (* An actual C compilation wants to drop these two. *)
-    Options.drop := [
-      Bundle.Module [ "FStar"; "Int"; "Cast"; "Full" ];
-      Bundle.Module [ "C" ];
-      Bundle.Module [ "TestLib" ]
-    ] @ !Options.drop
-  else begin
+  if !Options.wasm then begin
     (* True Wasm compilation: this module is useless (only assume val's). *)
     (* Only keep what's stricly needed from the C module. *)
-    Options.bundle := ([], [ Bundle.Module [ "C" ]]) :: !Options.bundle
+    Options.bundle := ([], [ Bundle.Module [ "C" ]]) :: !Options.bundle;
+    Options.extract_uint128 := true
   end;
 
   (* Self-help. *)
@@ -328,16 +326,17 @@ Supported options:|}
   if not !Options.minimal then
     Options.bundle :=
       ([], [ Bundle.Module [ "C"; "Loops" ]; Bundle.Module [ "Spec"; "Loops" ] ]) ::
+      ([], [ Bundle.Module [ "Prims" ] ]) ::
       ([], [ Bundle.Prefix [ "FStar" ] ]) ::
       ([], [ Bundle.Prefix [ "LowStar" ] ]) ::
       !Options.bundle;
 
   if !arg_c89 then begin
-    Options.uint128 := false;
     Options.anonymous_unions := false;
     Options.compound_literals := `Never;
     Options.c89_scope := true;
-    Options.c89_std := true
+    Options.c89_std := true;
+    Options.ccopts := Driver.Dash.d "KRML_VERIFIED_UINT128" :: !Options.ccopts
   end;
 
   (* Then, bring in the "default options" for each compiler. *)
@@ -353,13 +352,6 @@ Supported options:|}
     Warnings.(maybe_fatal_error ("", Deprecated ("-drop", "use a combination of \
       -bundle and -d reachability to make sure the functions are eliminated as \
       you wish")));
-
-  (* If the compiler supports uint128, then we just drop the module and let
-   * dependency analysis use FStar.UInt128.fsti. If the compiler does not,
-   * then we bring the implementation into scope instead. The latter is
-   * performed in src/Driver.ml because we need to know where FSTAR_HOME is. *)
-  if !Options.uint128 then
-    Options.drop := Bundle.Module [ "FStar"; "UInt128" ] :: !Options.drop;
 
   (* Timings. *)
   Time.start ();
@@ -400,53 +392,12 @@ Supported options:|}
   let filenames = List.filter (fun f -> Filename.basename f <> "prims.krml") filenames in
   let files = InputAst.read_files filenames in
 
-  (* These are modules we don't want because there's primitive support for them
-   * in kremlin. *)
-  let should_keep (name, _) = match name with
-    | "FStar_Int8" | "FStar_UInt8" | "FStar_Int16" | "FStar_UInt16"
-    | "FStar_Int31" | "FStar_UInt31" | "FStar_Int32" | "FStar_UInt32"
-    | "FStar_Int63" | "FStar_UInt63" | "FStar_Int64" | "FStar_UInt64"
-    | "FStar_Int128" | "FStar_HyperStack_ST" | "FStar_Monotonic_HyperHeap"
-    | "LowStar_Buffer" | "LowStar_Modifies"
-    | "FStar_Buffer" | "FStar_Monotonic_HyperStack" | "FStar_Monotonic_Heap"
-    | "C_String" | "FStar_Dyn" ->
-        false
-    | "FStar_UInt128" ->
-        (* Keep if we don't use the uint128 type. *)
-        not !Options.uint128
-    | "WasmSupport" ->
-        (* Keep if we're compiling to Wasm. *)
-        !Options.wasm
-    | _ ->
-        true
-  in
-  let files = List.filter should_keep files in
-
-  (* A quick sanity check to save myself some time. *)
-  begin try
-    let has_uint128 =
-      List.exists (function
-        | InputAst.DTypeFlat (([ "FStar"; "UInt128" ], "uint128"), _, _, _) ->
-            true
-        | _ ->
-            false
-      ) (List.assoc "FStar_UInt128" files)
-    in
-    (* The input file defines uint128 if and only if the backend does not
-     * support it. *)
-    if has_uint128 <> not !Options.uint128 then
-      Warnings.fatal_error "The implementation of FStar.UInt128 should be \
-        present in the input when using -fnouint128."
-  with Not_found ->
-    ()
-  end;
-
   (* -djson *)
   if !arg_print_json then
     Yojson.Safe.to_channel stdout (InputAst.binary_format_to_yojson (InputAst.current_version, files));
 
   (* -dast *)
-  let files = Builtin.prelude () @ Builtin.augment (InputAstToAst.mk_files files) in
+  let files = Builtin.prepare (InputAstToAst.mk_files files) in
   if !arg_print_ast then
     print PrintAst.print_files files;
 
@@ -456,6 +407,7 @@ Supported options:|}
    * - A needs to come before B so that in the resulting bundle, "static void
    *   A_f" comes before "static void B_g" (since they're static, there's no
    *   forward declaration in the header. *)
+  let files = Builtin.make_libraries files in
   let files = Bundles.topological_sort files in
 
   (* 1. We create bundles, and monomorphize functions first. This creates more
@@ -541,7 +493,6 @@ Supported options:|}
   let files = if !Options.tail_calls then Simplify.tail_calls files else files in
   let files = Simplify.simplify2 files in
   let files = Inlining.cross_call_analysis files in
-  let files = if !Options.wasm then SimplifyWasm.compile_copy_assignments files else files in
   if !arg_print_structs then
     print PrintAst.print_files files;
   let has_errors, files = Checker.check_everything files in